Make /learning the course list; move resource hub to /learning/resources
- /learning now renders the course list (image-left/content-right cards)
- Append a hardcoded "Guides & Resources" card linking to /learning/resources
- Move the DevResource hub grid from /learning to /learning/resources
- Point the hub's "Courses" feature card to /learning
- Remove the now-redundant /learning/courses list page (detail pages stay at /learning/courses/[slug])
- Redirect /learn -> /learning (course list); /learn/:path* -> /learning/courses/:path*
